Refactor Otto 2 auth to Web Core specific structure (#1784)
Moved middleware and strategies to apps/web/core (application-specific)
and simplified error handling for Vue SPA architecture.

Changes:
- Move lib/middleware/* → apps/web/core/middleware/*
- Move lib/onetime/web_auth_strategies.rb → apps/web/core/auth_strategies.rb
- Simplify error handling middleware (Vue handles errors client-side)
- Update module namespaces (Middleware → Core::Middleware)
- Update class names (OnetimeWebXStrategy → XStrategy)

Architecture rationale:
- Web Core serves Vue SPA entry points, not full HTML error pages
- V2 API handles all data operations
- Error middleware simplified to serve Vue entry point with status codes
- Removed complex exception handling (FormError, RecordNotFound, etc.)
- Let Vue application handle error display and user feedback

Files:
- apps/web/core/auth_strategies.rb (118 lines)
- apps/web/core/middleware/error_handling.rb (119 lines, down from 300+)
- apps/web/core/middleware/request_setup.rb (87 lines)
- apps/middleware_stack.rb (updated require paths)
- apps/web/core/app.rb (updated require paths)
